UE5->UE4的变量替换

+++ Unreal/CarlaUE4/Plugins/UnrealRoboticsLab/Source/URLab/Public/MuJoCo/Components/Actuators/MjIntVelocityActuator.h
@@ -51,7 +51,7 @@ public:
 
     /** @brief Derivative gain (Kv) for velocity damping. */
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "MuJoCo|Actuator|Parameters", meta=(EditCondition="bOverride_Kv"))
-    double Kv = 0.0f;
+    float Kv = 0.0f;
 
     /** @brief Override toggle for DampRatio. */
     UPROPERTY(EditAnywhere, Category = "MuJoCo|Actuator", meta=(InlineEditConditionToggle))
